Intento arreglo bug

+++ EPD3-2/app/Http/Controllers/tarjetaController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 use App\Models\User;
 use App\Models\CreditCard;
+use App\Models\Order;
+use App\Models\Ticket;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
@@ -19,7 +21,7 @@ class tarjetaController extends Controller
     {
         $user = User::where('id', Auth::id())->first();
         $tarjeta = CreditCard::where('users_id',$user->id)->first();
-        return view('tarjeta_read', compact('tarjeta')); 
+        return view('tarjeta_read', compact('tarjeta'));
     }
 
     /**
@@ -54,9 +56,9 @@ class tarjetaController extends Controller
             'fecha_caducidad.required' => 'El campo Fecha de caducidad es obligatorio y debe tener formato mes/año.',
             'cvc.required' => 'El campo CVC es obligatorio y debe tener 3 dígitos.',
         ];
-    
+
         $validator = Validator::make($request->all(), $rules, $messages);
-    
+
         if ($validator->fails()) {
             return redirect()->route('creditCard.create')->withErrors($validator)->withInput();
         }
@@ -67,7 +69,7 @@ class tarjetaController extends Controller
         $tarjeta->fecha_caducidad = $request->fecha_caducidad;
         $tarjeta->CVC = $request->cvc;
         $tarjeta->save();
-        return redirect('/home')->with('mensaje', 'La tarjeta se ha creado correctamente');
+        return redirect('/home')->with('success-perfil', 'La tarjeta se ha creado correctamente');
     }
 
     /**
@@ -112,9 +114,9 @@ class tarjetaController extends Controller
             'fecha_caducidad.required' => 'El campo Fecha de caducidad es obligatorio y debe tener formato mes/año.',
             'cvc.required' => 'El campo CVC es obligatorio y debe tener 3 dígitos.',
         ];
-    
+
         $validator = Validator::make($request->all(), $rules, $messages);
-    
+
         if ($validator->fails()) {
             return redirect()->route('creditCard.edit', $tarjeta)->withErrors($validator)->withInput();
         }
@@ -136,7 +138,29 @@ class tarjetaController extends Controller
     {
         $user = User::where('id', Auth::id())->first();
         $tarjeta = CreditCard::where('users_id',$user->id)->first();
-        $tarjeta->delete();
-        return redirect('/home')->with('mensaje', 'La tarjeta se ha eliminado correctamente.');
+        $pedidos = Order::where('credit_cards_id',$tarjeta->id)->get();
+        $puede=True;
+        foreach ($pedidos as $pedido ) {
+            if($pedido->state !='entregado'){
+                $puede = False;
+            }else{
+                $tickets = Ticket::where('orders_id', $pedido->id)->get();
+
+                // Borrar los registros de phones
+                foreach ($tickets as $ticket) {
+                    $ticket->delete();
+                }
+
+                $pedido->delete();
+            }
+        }
+
+        if($puede){
+            $tarjeta->delete();
+            return redirect('/home')->with('success-perfil', 'La tarjeta se ha eliminado correctamente.');
+        }else{
+             return redirect('/home')->with('success-perfil', 'La tarjeta no se ha eliminado correctamente.');
+        }
+
     }
 }
